#sdy Fix forward of making JAX changes so we can fall back to GSPMD in JAX export if the loaded module was lowered for GSPMD.
The final module that will be created by JAX export will contain a bit of Shardy and GSPMD ops. What we then do during compilation is detect whether there is a mix of these ops. If there is, we override the build option and instead use GSPMD for propagation (we have well tested code to export Shardy->GSPMD, but not vice versa).

Also fix issue in `hasGspmdAttrsOrOps` helper by only checking arg/output shardings of the non-main function. This fixes an issue where for shape polymorphism, it inlines the functions, meaning we need to walk the body of the main function. However we still don't need to check the inputs/outputs of the main function, since they are guaranteed to have Shardy shardings.

Reverts 92d5fe88d0a2cd9a8c84bdc2fab85f2995f644b6

PiperOrigin-RevId: 778831460