[pallas] Forked `load` and `store` into `triton` and `tpu`
The thinking here is that these ops are fundamentally backend-specific,
and thus each backend should (and does in practice) have its own flavor.

This change does not introduce any new primitives, but it does restrict
the signature of `pltpu` variants to the parameters which make sense on TPU.

I will deprecate `pl.load` and `pl.store` in a follow up.

PiperOrigin-RevId: 785852974